plugin/colorization/requirements.txt
torch
torchvision
numpy
opencv-python
opencv-transforms
pillow